% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compmDistMat.R
\name{compmDistMat}
\alias{compmDistMat}
\title{Distance matrix among \eqn{m}-dimensional functions / trajectories}
\usage{
compmDistMat(
  x,
  y = NULL,
  method = "Euclidean",
  measure = NULL,
  timestamps = NULL,
  reset_timestamps = FALSE,
  resample = TRUE,
  flip_threshold = 0,
  hover_threshold = 2000,
  baseline = FALSE,
  weight = NULL,
  parallel = FALSE,
  cl = NULL,
  diag = TRUE,
  upper = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{list} of \eqn{m} objects in matrix form. Each matrix stores a dimension of the set of functions, such that columns are individuals (\eqn{n}) and rows are discrete-time points (\eqn{t}). Functions' values should be of the same time points.}

\item{y}{\code{list} of \eqn{m} objects in matrix form. Each matrix stores a dimension of the set of functions, such that columns are individuals (\eqn{k}) and rows are discrete-time points (\eqn{t}). Functions' values should be of the same time points.}

\item{method}{\code{string} to specify the distance measure to compute. Defaults to "Euclidean". Can be set to: ...}

\item{measure}{\code{string} to specify the measure to be computed with the  mousetrap package (can only be used instead of \code{"method"}).
The complete list of available measures can be retrieved with the function \code{\link[compmDistMat:mmeasuresChoices]{mmeasuresChoices}}.
For a detailed explanation of the different measures see \code{\link[mousetrap:mt_measures]{mousetrap:mt_measures}}. Can only be used for 2D data.}

\item{baseline}{is a binary variable indicating whether the distance matrix should be corrected by the baseline characteristics}

\item{weight}{OTHER PARAMETERS OF MOUSETRAP TO BE FINISHED}

\item{parallel}{logical value indicating whether computations should be parallelized. Default is FALSE. If TRUE, parallelization is conducted with \href{https://www.rdocumentation.org/packages/parallel}{parallel} package.}

\item{cl}{a cluster object created by \href{https://www.rdocumentation.org/packages/parallel}{parallel}. Default is NULL.}

\item{diag}{is a logical value indicating whether the diagonal of the distance matrix should be printed. Only applicable if distances of one set of functions with itself is computed (\code{isnull(y)=TRUE}). If TRUE, diagonal values are NA.}

\item{upper}{is a logical value indicating whether the upper triangle of the matrix should be printed. Only applicable if distances of one set of functions with itself is computed (\code{isnull(y)=TRUE}). If TRUE, upper triangle values are NA.}
}
\value{
The present function returns a rectangular matrix of size \eqn{k*n} containing the distances of
all \eqn{k} functions in \code{y} with all n functions/trajectories in \code{x}. If \code{y} is left unspecified, the function returns the distances of the functions in \code{x} with itself (a symmetric \eqn{n x n} matrix).
}
\description{
The present function computes the distance matrix of one set of \eqn{n} different \eqn{m}-dimensional functions to another set of \eqn{k} different \eqn{m}-dimensional functions. The distances for all combinations of functions in the first set with functions of the second set are returned.
Different distance measures are available and can be set by \code{method}.
}
\examples{
## 2-dimensional functions

xdim1 <- replicate(4, rnorm(100, 0, 3))
xdim2 <- replicate(4, rnorm(100, 3, 1))

ydim1 <- replicate(2, rnorm(100, 0, 3))
ydim2 <- replicate(2, rnorm(100, 3, 1))

x <- list(xdim1, xdim2)
y <- list(ydim1, ydim2)
compmDistMat(x = x, y = y, method = "Euclidean", parallel = FALSE, cl = NULL)

## 3-dimensional functions
xdim1 <- replicate(4, rnorm(100, 0, 3))
xdim2 <- replicate(4, rnorm(100, 3, 1))
xdim3 <- replicate(4, rnorm(100, 5, 1))

ydim1 <- replicate(2, rnorm(100, 0, 3))
ydim2 <- replicate(2, rnorm(100, 3, 1))
ydim3 <- replicate(2, rnorm(100, 5, 1))

x <- list(xdim1, xdim2, xdim3)
y <- list(ydim1, ydim2, ydim3)
compmDistMat(x = x, y = y, method = "Euclidean", parallel = FALSE, cl = NULL)

compmDistMat: computes the distance (given "method") among n-dimensional vector(s)

x is a list object with the components/coordinates (e.g., as matrices with columns as functions) of the n-dimensional functions (to fit the model)
y is a list object with the components/coordinates (e.g., as matrices with columns as functions) of the n-dimensional functions (to predict)
method is the distance measure to be used (similar to stat:dist)
measure is a logical value (defaul=FALSE). If TRUE, mtmeasures() function is called and measures are computed with mousetrap package
baseline is a binary variable indicating whether the distance matrix should be corrected by the baseline characteristics
diag is a logical value indicating whether the diagonal of the distance matrix should be printed (similar to stat:dist)
upper is a logical value indicating whether the upper triangle of the matrix should be printed (similar to stat:dist)
}
\seealso{
See \code{\link[parallel:makeCluster]{makeCluster}}, \code{\link[parallel:clusterExport]{clusterExport}}, \code{\link[parallel:stopCluster]{stopCluster}}, \code{\link[parallel:parApply]{parApply}} and \code{\link[parallel:parLapply]{parLapply}} from  \href{https://www.rdocumentation.org/packages/parallel}{parallel}
}
