% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiKernel-mlr.R
\name{makeRLearner.classif.mclassiKernel}
\alias{makeRLearner.classif.mclassiKernel}
\title{MLR Implementation Kernel
makeRLearner.classif.mclassiKernel}
\usage{
\method{makeRLearner}{classif.mclassiKernel}()
}
\description{
MLR Implementation Kernel
makeRLearner.classif.mclassiKernel
}
