% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiKnn-mlr.R
\name{makeRLearner.classif.mclassiKnn}
\alias{makeRLearner.classif.mclassiKnn}
\title{MLR Implementation KNN
makeRLearner.classif.mclassiKnn}
\usage{
\method{makeRLearner}{classif.mclassiKnn}()
}
\description{
MLR Implementation KNN
makeRLearner.classif.mclassiKnn
}
