% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_char_sequences_for_malignmentscores.R,
%   R/malignmentscores.R
\name{malignmentscore}
\alias{malignmentscore}
\title{Sequence alignment scores between multivariate functions / trajectories}
\usage{
malignmentscore(data, Keeprep = TRUE, parallel = FALSE, cl = NULL, ...)

malignmentscore(data, Keeprep = TRUE, parallel = FALSE, cl = NULL, ...)
}
\arguments{
\item{data}{a list of \eqn{m} objects in matrix form. Each matrix stores a dimension of the set of functions, such that columns are individuals (\eqn{n}) and rows are discrete-time points (\eqn{t}). Functions' values should be of the same time points.}

\item{Keeprep}{logical value indicating whether adjacent duplicates should be kept or removed.
For instance, if the function / trajectory has values within the same area for consecutive time stamps, should those consecutive duplicates be kept? Default is TRUE.}

\item{parallel}{logical value indicating whether computations should be parallelized. Default is FALSE. If TRUE, parallelization is conducted with \href{https://www.rdocumentation.org/packages/parallel}{parallel} package.}

\item{cl}{a cluster object created by \href{https://www.rdocumentation.org/packages/parallel}{parallel}. Default is NULL.}

\item{...}{additional arguments that can be passed to \code{\link[Biostrings:pairwiseAlignment]{Biostrings::pairwiseAlignment}}}

\item{dim_min_max}{list of size \eqn{m} (number of dimensions), containing vectors of size 2 giving the minimum and maximum values of the functions in each dimensions.}

\item{n_breaks}{vector of integers giving the number of breaks in each dimension. Limited to max 10 per dimension and max 500 areas in total.}

\item{custom_splits}{list of size \eqn{m} (number of dimensions of the functions) containing vectors giving the exact value at which to include splits in each dimension. In the format list(c(lower bound, break1, break2,..., breakn, upperbound), ...). Either define \code{custom_splits} or (\code{dim_min_max} and \code{n_breaks})}
}
\value{
Returns a square and symmetric \eqn{n x n} matrix of alignment scores.
}
\description{
Computes Sequence Alignment scores for all pairs of \eqn{m}-dimensional functions.
Sequence alignment scores are usually computed for aligning sequences of nominal values. The present function extends

Computes Sequence Alignment scores for all pairs of \eqn{m}-dimensional functions.
Sequence alignment scores are usually computed for aligning sequences of nominal values. The present function extends
the computation of such alignment scores to functions by splitting the function space into areas.
Thus, alignment scores of the sequence of locations can be calculated.
}
\details{
The present function translates each coordinate at each time point into an area.
The sequence of areas (encoded as characters) is then used in the \code{\link[Biostrings:pairwiseAlignment]{Biostrings::pairwiseAlignment}}
function.
}
\seealso{
See \code{\link[parallel:makeCluster]{makeCluster}}, \code{\link[parallel:clusterExport]{clusterExport}}, \code{\link[parallel:stopCluster]{stopCluster}}, \code{\link[parallel:parApply]{parApply}} and \code{\link[parallel:parLapply]{parLapply}} from  \href{https://www.rdocumentation.org/packages/parallel}{parallel}
}
