% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiKernel.R
\name{mclassiKernel}
\alias{mclassiKernel}
\title{mclassiKernel:}
\usage{
mclassiKernel(
  classes,
  fdata,
  metric = NULL,
  nderiv = 0L,
  kernel = "Ker.norm",
  h = 1,
  baseline = FALSE,
  weight = NULL,
  measure = NULL,
  timestamps = NULL,
  reset_timestamps = FALSE,
  resample = TRUE,
  flip_threshold = 0,
  hover_threshold = 2000,
  parallel = FALSE,
  cl = NULL,
  diag = FALSE,
  upper = FALSE,
  ...
)
}
\description{
mclassiKernel: computes the distance (given "method") among n-dimensional vector(s) #
}
