% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiKnn.R
\name{mclassiKnn}
\alias{mclassiKnn}
\title{Create mKNN Object}
\usage{
mclassiKnn(
  classes,
  fdata,
  mdist = NULL,
  metric = "Euclidean",
  knn = 1L,
  nderiv = 0L,
  baseline = FALSE,
  weight = NULL,
  measure = NULL,
  timestamps = NULL,
  reset_timestamps = FALSE,
  resample = TRUE,
  flip_threshold = 0,
  hover_threshold = 2000,
  parallel = FALSE,
  cl = NULL,
  diag = FALSE,
  upper = FALSE,
  ...
)
}
\arguments{
\item{classes}{\code{factor} or \code{numeric}. A vector containing the true classes of the training data.}

\item{fdata}{the training covariates as a \code{list} of \eqn{m} objects in \code{matrix} form. Each matrix stores a dimension of the set of functions, such that columns are individuals (\eqn{n}) and rows are discrete-time points (\eqn{t}). Functions' values should be of the same time points.}

\item{metric}{\code{string} specifying the metric for knn. The complete list of available metrics is returned by
\code{mdistMeasures()}}

\item{knn}{\code{integer} specifying the number of nearest neighbours considered in the knn algorithm.}

\item{nderiv}{\code{integer} number of derivatives of the trajectory.}

\item{baseline}{is a binary variable indicating whether the distance matrix should be corrected by the baseline characteristics}

\item{measure}{is a logical value (defaul=FALSE). If TRUE, mtmeasures() function is called and measures are computed with mousetrap package}

\item{parallel}{logical value indicating whether computations should be parallelized. Default is FALSE. If TRUE, parallelization is conducted with \href{https://www.rdocumentation.org/packages/parallel}{parallel} package.}

\item{cl}{a cluster object created by \href{https://www.rdocumentation.org/packages/parallel}{parallel}. Default is NULL.}

\item{diag}{is a logical value indicating whether the diagonal of the distance matrix should be printed. Only applicable if distances of one set of functions with itself is computed (\code{isnull(y)=TRUE}). If TRUE, diagonal values are NA.}

\item{upper}{is a logical value indicating whether the upper triangle of the matrix should be printed. Only applicable if distances of one set of functions with itself is computed (\code{isnull(y)=TRUE}). If TRUE, upper triangle values are NA.}
}
\value{
Returns a list containing the parameters and the training data for a mknn prediction. This list is used for prediction with the function \code{\link[classiMultiFuncTest:predict.mclassiKnn]{predict.mclassiKnn}}
}
\description{
The present function creates an object to specify the k-nearest-neighbor classification
for \eqn{n}-different observations of \eqn{m}-dimensional functionnal data.
}
\examples{

## 3-dimensional functions
# classes
classes <- as.factor(c("Cat", "Dog", "Dog", "Cat"))

# fdata
x <- replicate(4, rnorm(100, 0, 3))
y <- replicate(4, rnorm(100, 3, 1))
z <- replicate(4, rpois(100, 2))
fdata <- list(x, y, z)


mclassiKnn(classes = classes, fdata = fdata, knn = 3, nderiv = 0, cl = NULL)
}
\seealso{
See \code{\link[parallel:makeCluster]{makeCluster}}, \code{\link[parallel:clusterExport]{clusterExport}}, \code{\link[parallel:stopCluster]{stopCluster}}, \code{\link[parallel:parApply]{parApply}} and \code{\link[parallel:parLapply]{parLapply}} from  \href{https://www.rdocumentation.org/packages/parallel}{parallel}
}
