% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiNestedCV.R
\name{mclassiOnested}
\alias{mclassiOnested}
\title{mclassiOnested}
\usage{
mclassiOnested(
  cl,
  task,
  knn = NULL,
  kernel = NULL,
  id,
  par.vals,
  hyperparams.vals = NULL,
  super.learner = "boosting",
  cv = 5,
  M = 100,
  weight
)
}
\description{
task: define a task with makeClassifTask
data: data set with: id, dimension, trajectories as rows and condition
target: condition
knn: value for nearest neighbor
par.vals: hyperparameters
cv: nested loop (model performance) is done given k-fold cross-validation (cv=k)
M: number of iterations in the inner loop (subsampling)
weight: subsampling is used to tuned parameters. "weight" is a vector with weights for training and testing data sets
}
