% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiNestedCV.R
\name{mclassiPerf}
\alias{mclassiPerf}
\title{mclassiPerf}
\usage{
mclassiPerf(
  cl,
  task,
  id,
  knn = NULL,
  kernel = NULL,
  par.vals,
  subset,
  super.learner = NULL,
  par.super.learner = NULL
)
}
\description{
cl: classifier
task: define a task with makeClassifTask
data: data set with: id, dimension, trajectories as rows and condition
target: condition
knn: value for nearest neighbor
kernel: type of kernel to be used
par.vals: hyperparameters
subset: ith inner sample
#'@export
}
