% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiEnsemble-mlr.R
\name{mclassiStackLearner}
\alias{mclassiStackLearner}
\title{mclassiEnsemble}
\usage{
mclassiStackLearner(task, cl, id, knn, kernel = NULL, par.vals, subset = NULL)
}
\description{
mclassiEnsemble
}
