% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclassiTies.R
\name{mclassiTies}
\alias{mclassiTies}
\title{mclassiTies}
\usage{
mclassiTies(x, knn, ...)
}
\arguments{
\item{x}{vector of distances of a new trajectory to all training trajectories.}

\item{knn}{number of nearest neighbours}
}
\description{
Provides the positions of the indices of the k nearest trajectories.
Selects randomly among ties for the \eqn{k}th nearest trajectory.
}
\examples{
## distance of new data to the 10 trajectories in the training data:
dist_to_train_fucntions <- c(1, 2, 2, 2, 2, 2, 5, 5, 5, 6)
mclassiTies(x = dist_to_train_fucntions, knn = 3)
}
