% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdistMeasures.R
\name{mdistMeasures}
\alias{mdistMeasures}
\title{Distance between measures of 2-Dimensional trajectories.}
\usage{
mdistMeasures(data, measure)
}
\arguments{
\item{data}{\code{list} containing one vector (or two if \code{measure="flips"})
of measures computed by the function \code{\link[classiMultiFunc:mmeasures]{mmeasures}}}

\item{measure}{\code{string} to specify the measure contained in \code{data}.
The complete list of available measures can be retrieved with the function \code{\link[compmDistMat:mmeasuresChoices]{mmeasuresChoices}}.
For a detailed explanation of the different measures see \code{\link[mousetrap:mt_measures]{mousetrap:mt_measures}}.}
}
\description{
The mousetrap package enables the calculation of certain measures of mouse trajectories.
The present function enables the computation of "distances" between such measures.
}
\examples{
#Example 1: measure != "flips"
data = list(c(1,2,3))
mdistMeasures(data=data, measure ="ypos_max")

#Example 2: measure = "flips"
data = list(c(1,2,3), c(4,5,6))
mdistMeasures(data=data, measure ="flips")

return
}
