% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdtw.R
\name{mdtw}
\alias{mdtw}
\title{Dynamic Time Warping distance between multivariate functions}
\usage{
mdtw(data, parallel = FALSE, cl = NULL, ...)
}
\arguments{
\item{data}{a list of \eqn{m} objects in matrix form. Each matrix stores a dimension of the set of functions, such that columns are individuals (\eqn{n}) and rows are discrete-time points (\eqn{t}). Functions' values should be of the same time points.}

\item{parallel}{logical value indicating whether computations should be parallelized. Default is FALSE. If TRUE, parallelization is conducted with \href{https://www.rdocumentation.org/packages/parallel}{parallel} package.}

\item{cl}{a cluster object created by \href{https://www.rdocumentation.org/packages/parallel}{parallel}. Default is NULL.}

\item{...}{additional arguments that can be passed to \code{\link[dtw:dtw]{dtw::dtw}}}
}
\value{
Returns a square and symmetric \eqn{n x n} matrix of \eqn{m}-dimensional Dynamic Time Warp distances.
}
\description{
Computes the Dynamic Time Warping distance for all pairs of \eqn{m}-dimensional functions.
}
\details{
Applies the \code{\link[dtw:dtw]{dtw::dtw}} function  to all pairs of \eqn{m}-dimensional functions.
By default, the Euclidean Distance between optimally aligned \eqn{m}-dimensional functions is computed.
Other options and parameters that can be passed to \code{\link[dtw:dtw]{dtw::dtw}} (such as the use of the Manhattan distance) can be defined in the additional parameters \code{...}
}
\examples{
## 2-dimensional functions

x = replicate(4, rnorm(100, 0, 3))
y = replicate(4, rnorm(100, 3, 1))
data = list(x, y)
mdtw(data, parallel = FALSE, cl = NULL)

## 3-dimensional functions

z = replicate(4, rpois(100, 2))
data = list(x, y, z)
mdtw(data, parallel = FALSE, cl = NULL)

}
\seealso{
See \code{\link[parallel:makeCluster]} \code{\link[parallel:makeCluster]{makeCluster}}, \code{\link[parallel:clusterExport]{clusterExport}}, \code{\link[parallel:stopCluster]{stopCluster}}, \code{\link[parallel:parApply]{parApply}} and \code{\link[parallel:parLapply]{parLapply}} from  \href{https://www.rdocumentation.org/packages/parallel}{parallel}, \code{\link[proxy:dist]{dist}} from \href{https://cran.r-project.org/web/packages/proxy/index.html}{proxy}
}
