% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mglobMin_2_3.R
\name{mglobMin3}
\alias{mglobMin3}
\title{Global Minimum distance between multivariate functions}
\usage{
mglobMin3(data, parallel = FALSE, cl = NULL)
}
\arguments{
\item{data}{a matrix that stores a dimension of the set of functions, such that columns are individuals (\eqn{n}) and rows are discrete-time points (\eqn{t}). Functions' values should be of the same time points.}

\item{parallel}{logical value indicating whether computations should be parallelized. Default is FALSE. If TRUE, parallelization is conducted with \href{https://www.rdocumentation.org/packages/parallel}{parallel} package.}

\item{cl}{a cluster object created by \href{https://www.rdocumentation.org/packages/parallel}{parallel}. Default is NULL.}
}
\value{
Returns a square and symmetric \eqn{n x n} matrix of \eqn{m}-dimensional global minimum distances.
}
\description{
Computes the Global Minimum distance for all pairs of \eqn{m}-dimensional functions. For a single pair of functions,
the present R function returns the minimum euclidean distance between the function values at equal time points.
}
\details{
For each pair of functions f and g, the present R function computes: \eqn{min t [Euclidean_Distance(f(t), g(t))]}
}
\examples{
## 2-dimensional functions

x = replicate(4, rnorm(100, 0, 3))
y = replicate(4, rnorm(100, 3, 1))
data = list(x, y)
mglobmin(data, parallel = FALSE, cl = NULL)

## 3-dimensional functions

z = replicate(4, rpois(100, 2))
data = list(x, y, z)
mglobmin(data, parallel = FALSE, cl = NULL)

}
\seealso{
See \code{\link[parallel:makeCluster]{makeCluster}}, \code{\link[parallel:clusterExport]{clusterExport}}, \code{\link[parallel:stopCluster]{stopCluster}}, \code{\link[parallel:parApply]{parApply}} and \code{\link[parallel:parLapply]{parLapply}} from  \href{https://www.rdocumentation.org/packages/parallel}{parallel}, and \code{\link[proxy:dist]{dist}} from \href{https://cran.r-project.org/web/packages/proxy/index.html}{proxy}
}
