% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmeasures.R
\name{mmeasures}
\alias{mmeasures}
\title{mmeasures: computes measures from mousetrap package}
\usage{
mmeasures(
  mtdata,
  timestamps,
  reset_timestamps = FALSE,
  resample = TRUE,
  flip_threshold = 0,
  hover_threshold = 2000
)
}
\description{
mtdata is a data frame with id, x- and y-coordinates (in this order)
timestamps is a data frame with the timestamps of the trajectories (training + testing order)
reset_timestamps is a logical value from mt_import_mousetrap (see moustrap package). Defaul FALSE
resample is a logical value. When TRUE (default) the function mt_resample is executed (see moustrap package)
flip_threshold is a numerical value from mt_measures (see moustrap package)
hover_threshold is a numerical value from mt_measures (see moustrap package)
}
