% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpoints.R
\name{mpoints}
\alias{mpoints}
\title{Mean distance between positions at pre-defined time points}
\usage{
mpoints(data, tquantiles, parallel = FALSE, cl = NULL)
}
\arguments{
\item{data}{a \code{list} of \eqn{m} objects in matrix form. Each matrix stores a dimension of the set of functions, such that columns are individuals (\eqn{n}) and rows are discrete-time points (\eqn{t}). Functions' values should be of the same time points.}

\item{tquantiles}{\code{vector} of time stamp quantiles at which distances between curves should be measured.}

\item{parallel}{logical value indicating whether computations should be parallelized. Default is FALSE. If TRUE, parallelization is conducted with \href{https://www.rdocumentation.org/packages/parallel}{parallel} package.}

\item{cl}{a cluster object created by \href{https://www.rdocumentation.org/packages/parallel}{parallel}. Default is NULL.}
}
\value{
Returns a square and symmetric \eqn{n x n} matrix of \eqn{m}-dimensional "Jump" differences.
}
\description{
For each \eqn{m}-dimensional function/trajectory the present R function computes the mean distance between curves at pre-defined time points.
}
\examples{
## 2-dimensional functions

x = replicate(4, rnorm(100, 0, 3))
y = replicate(4, rnorm(100, 3, 1))
data = list(x, y)
mpoints(data,tquantiles=c(0, 0.2, 0.8) , parallel = FALSE, cl = NULL)

## 3-dimensional functions

z = replicate(4, rpois(100, 2))
data = list(x, y, z)
mpoints(data,tquantiles=c(0, 0.2, 0.8) , parallel = FALSE, cl = NULL)

}
\seealso{
See \code{\link[parallel:makeCluster]{makeCluster}}, \code{\link[parallel:clusterExport]{clusterExport}}, \code{\link[parallel:stopCluster]{stopCluster}}, \code{\link[parallel:parApply]{parApply}} and \code{\link[parallel:parLapply]{parLapply}} from  \href{https://www.rdocumentation.org/packages/parallel}{parallel}, and \code{\link[proxy:dist]{dist}} from \href{https://cran.r-project.org/web/packages/proxy/index.html}{proxy}
}
