% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mclassiKernel.R
\name{predict.mclassiKernel}
\alias{predict.mclassiKernel}
\title{predict_mclassiKernel}
\usage{
\method{predict}{mclassiKernel}(object, newdata = NULL, predict.type = "response")
}
\description{
predict_mclassiKernel
}
